実績のクラス呼び出しをconstファイルに切り出し
